+++ src/components/InventoryTable/EntityTableToolbar.js
@@ -19,7 +19,9 @@ import {
     TAG_CHIP,
     arrayToSelection,
     RHCD_FILTER_KEY,
-    UPDATE_METHOD_KEY
+    UPDATE_METHOD_KEY,
+    LAST_SEEN_CHIP,
+    HOST_GROUP_CHIP
 } from '../../Utilities/index';
 import { onDeleteFilter, onDeleteTag } from './helpers';
 import {
@@ -28,6 +30,7 @@ import {
     useRegisteredWithFilter,
     useTagsFilter,
     useRhcdFilter,
+    useLastSeenFilter,
     useUpdateMethodFilter,
     textFilterState,
     textFilterReducer,
@@ -40,9 +43,16 @@ import {
     rhcdFilterReducer,
     rhcdFilterState,
     updateMethodFilterReducer,
-    updateMethodFilterState
+    updateMethodFilterState,
+    lastSeenFilterReducer,
+    lastSeenFilterState,
+    groupFilterReducer,
+    groupFilterState
 } from '../filters';
 import useOperatingSystemFilter from '../filters/useOperatingSystemFilter';
+import useFeatureFlag from '../../Utilities/useFeatureFlag';
+import useGroupFilter from '../filters/useGroupFilter';
+import { DatePicker, Split, SplitItem } from '@patternfly/react-core';
 
 /**
  * Table toolbar used at top of inventory table.
@@ -80,14 +90,18 @@ const EntityTableToolbar = ({
         tagsFilterReducer,
         operatingSystemFilterReducer,
         rhcdFilterReducer,
-        updateMethodFilterReducer
+        lastSeenFilterReducer,
+        updateMethodFilterReducer,
+        groupFilterReducer
     ]), {
         ...textFilterState,
         ...stalenessFilterState,
         ...registeredWithFilterState,
         ...tagsFilterState,
         ...rhcdFilterState,
-        ...updateMethodFilterState
+        ...updateMethodFilterState,
+        ...lastSeenFilterState,
+        ...groupFilterState
     });
     const filters = useSelector(({ entities: { activeFilters } }) => activeFilters);
     const allTagsLoaded = useSelector(({ entities: { allTagsLoaded } }) => allTagsLoaded);
@@ -97,9 +111,14 @@ const EntityTableToolbar = ({
     const [stalenessFilter, stalenessChip, staleFilter, setStaleFilter] = useStalenessFilter(reducer);
     const [registeredFilter, registeredChip, registeredWithFilter, setRegisteredWithFilter] = useRegisteredWithFilter(reducer);
     const [rhcdFilterConfig, rhcdFilterChips, rhcdFilterValue, setRhcdFilterValue] = useRhcdFilter(reducer);
+    const [lastSeenFilter, lastSeenChip, lastSeenFilterValue, setLastSeenFilterValue,
+        toValidator, onFromChange, onToChange, endDate, startDate, rangeValidator] = useLastSeenFilter(reducer);
     const [osFilterConfig, osFilterChips, osFilterValue, setOsFilterValue] = useOperatingSystemFilter();
     const [updateMethodConfig, updateMethodChips, updateMethodValue, setUpdateMethodValue] = useUpdateMethodFilter(reducer);
+    const [hostGroupConfig, hostGroupChips, hostGroupValue, setHostGroupValue] = useGroupFilter();
 
+    const isUpdateMethodEnabled = useFeatureFlag('hbi.ui.system-update-method');
+    const groupsEnabled = useFeatureFlag('hbi.ui.inventory-groups');
     const {
         tagsFilter,
         tagsChip,
@@ -115,7 +134,7 @@ const EntityTableToolbar = ({
     const debounceGetAllTags = useCallback(debounce((config, options) => {
         if (showTags && !hasItems && hasAccess) {
             dispatch(fetchAllTags(config, {
-                ...options?.pagination
+                ...options?.paginationhideFilters
             },  getTags));
         }
     }, 800), [customFilters?.tags]);
@@ -127,7 +146,13 @@ const EntityTableToolbar = ({
         operatingSystem: !(hideFilters.all && hideFilters.operatingSystem !== false) && !hideFilters.operatingSystem,
         tags: !(hideFilters.all && hideFilters.tags !== false) && !hideFilters.tags,
         rhcdFilter: !(hideFilters.all && hideFilters.rhcdFilter !== false) && !hideFilters.rhcdFilter,
-        updateMethodFilter: !(hideFilters.all && hideFilters.updateMethodFilter !== false) && !hideFilters.updateMethodFilter
+        lastSeenFilter: !(hideFilters.all && hideFilters.lastSeen !== false) && !hideFilters.lastSeen,
+        //hides the filter untill API is ready. JIRA: RHIF-169
+        updateMethodFilter: isUpdateMethodEnabled &&
+            !(hideFilters.all && hideFilters.updateMethodFilter !== false)
+                && !hideFilters.updateMethodFilter,
+        hostGroupFilter: groupsEnabled && !(hideFilters.all && hideFilters.hostGroupFilter !== false)
+                && !hideFilters.hostGroupFilter
     };
 
     /**
@@ -165,7 +190,15 @@ const EntityTableToolbar = ({
      */
     useEffect(() => {
         const {
-            textFilter, tagFilters, staleFilter, registeredWithFilter, osFilter, rhcdFilter, updateMethodFilter
+            textFilter,
+            tagFilters,
+            staleFilter,
+            registeredWithFilter,
+            osFilter,
+            rhcdFilter,
+            lastSeenFilter,
+            updateMethodFilter,
+            groupFilter
         } = reduceFilters([...filters || [], ...customFilters?.filters || []]);
 
         debouncedRefresh();
@@ -176,6 +209,8 @@ const EntityTableToolbar = ({
         enabledFilters.operatingSystem && setOsFilterValue(osFilter);
         enabledFilters.rhcdFilter && setRhcdFilterValue(rhcdFilter);
         enabledFilters.updateMethodFilter && setUpdateMethodValue(updateMethodFilter);
+        enabledFilters.lastSeenFilter && setLastSeenFilterValue(lastSeenFilter);
+        enabledFilters.hostGroupFilter && setHostGroupValue(groupFilter);
     }, []);
 
     /**
@@ -256,10 +291,22 @@ const EntityTableToolbar = ({
     }, [rhcdFilterValue]);
 
     useEffect(() => {
+        if (shouldReload && enabledFilters.lastSeenFilter) {
+            onSetFilter(lastSeenFilterValue, 'lastSeenFilter', debouncedRefresh);
+        }
+    }, [lastSeenFilterValue]);
+
+    useEffect(() => {
         if (shouldReload && enabledFilters.updateMethodFilter) {
             onSetFilter(updateMethodValue, 'updateMethodFilter', debouncedRefresh);
         }
     }, [updateMethodValue]);
+
+    useEffect(() => {
+        if (shouldReload && enabledFilters.hostGroupFilter) {
+            onSetFilter(hostGroupValue, 'hostGroupFilter', debouncedRefresh);
+        }
+    }, [hostGroupValue]);
 
     /**
      * Mapper to simplify removing of any filter.
@@ -279,7 +326,9 @@ const EntityTableToolbar = ({
         ),
         [OS_CHIP]: (deleted) => setOsFilterValue(xor(osFilterValue, deleted.chips.map(({ value }) => value))),
         [RHCD_FILTER_KEY]: (deleted) => setRhcdFilterValue(onDeleteFilter(deleted, rhcdFilterValue)),
-        [UPDATE_METHOD_KEY]: (deleted) => setUpdateMethodValue(onDeleteFilter(deleted, updateMethodValue))
+        [LAST_SEEN_CHIP]: (deleted) => setLastSeenFilterValue(onDeleteFilter(deleted, [lastSeenFilterValue.mark])),
+        [UPDATE_METHOD_KEY]: (deleted) => setUpdateMethodValue(onDeleteFilter(deleted, updateMethodValue)),
+        [HOST_GROUP_CHIP]: (deleted) => setHostGroupValue(onDeleteFilter(deleted, hostGroupValue))
     };
     /**
      * Function to reset all filters with 'Reset Filter' is clicked
@@ -291,7 +340,9 @@ const EntityTableToolbar = ({
         enabledFilters.tags && setSelectedTags({});
         enabledFilters.operatingSystem && setOsFilterValue([]);
         enabledFilters.rhcdFilter && setRhcdFilterValue([]);
+        enabledFilters.lastSeenFilter && setLastSeenFilterValue([]);
         enabledFilters.updateMethodFilter && setUpdateMethodValue([]);
+        enabledFilters.hostGroupFilter && setHostGroupValue([]);
         dispatch(setFilter([]));
         updateData({ page: 1, filters: [] });
     };
@@ -310,6 +361,8 @@ const EntityTableToolbar = ({
                 ...!hasItems && enabledFilters.operatingSystem ? osFilterChips : [],
                 ...!hasItems && enabledFilters.rhcdFilter ? rhcdFilterChips : [],
                 ...!hasItems && enabledFilters.updateMethodFilter ? updateMethodChips : [],
+                ...!hasItems && enabledFilters.lastSeenFilter ? lastSeenChip : [],
+                ...!hasItems && enabledFilters.hostGroupFilter ? hostGroupChips : [],
                 ...activeFiltersConfig?.filters || []
             ],
             onDelete: (e, [deleted, ...restDeleted], isAll) => {
@@ -335,6 +388,8 @@ const EntityTableToolbar = ({
             ...enabledFilters.registeredWith ? [registeredFilter] : [],
             ...enabledFilters.rhcdFilter ? [rhcdFilterConfig] : [],
             ...enabledFilters.updateMethodFilter ? [updateMethodConfig] : [],
+            ...enabledFilters.lastSeenFilter ? [lastSeenFilter] : [],
+            ...enabledFilters.hostGroupFilter ? [hostGroupConfig] : [],
             ...showTags && enabledFilters.tags ? [tagsFilter] : []
         ] : [],
         ...filterConfig?.items || []
@@ -378,7 +433,31 @@ const EntityTableToolbar = ({
                 ...paginationProps
             } : <Skeleton size={SkeletonSize.lg} />}
         >
+            {lastSeenFilterValue?.mark === 'custom' &&
+            <Split>
+                <SplitItem>
+                    <DatePicker
+                        onChange={onFromChange}
+                        aria-label="Start date"
+                        validators={[rangeValidator]}
+
+                    />
+                </SplitItem>
+                <SplitItem style={{ padding: '6px 12px 0 12px' }}>
+            to
+                </SplitItem>
+                <SplitItem>
+                    <DatePicker
+                        value={endDate}
+                        onChange={onToChange}
+                        rangeStart={startDate}
+                        validators={[toValidator]}
+                        aria-label="End date"
+                    />
+                </SplitItem>
+            </Split>}
             { children }
+
         </PrimaryToolbar>
         {
             (showTags || enabledFilters.tags || showTagModal) && <TagsModal
@@ -424,7 +503,9 @@ EntityTableToolbar.propTypes = {
         stale: PropTypes.bool,
         operatingSystem: PropTypes.bool,
         rhcdFilter: PropTypes.bool,
+        lastSeen: PropTypes.bool,
         updateMethodFilter: PropTypes.bool,
+        hostGroupFilter: PropTypes.bool,
         all: PropTypes.bool
     }),
     paginationProps: PropTypes.object,