+++ src/components/InventoryTable/EntityTableToolbar.test.js
@@ -12,6 +12,13 @@ import TitleColumn from './TitleColumn';
 import debounce from 'lodash/debounce';
 
 jest.mock('lodash/debounce');
+jest.mock('../../Utilities/useFeatureFlag');
+
+jest.mock('../../Utilities/constants', () => ({
+    ...jest.requireActual('../../Utilities/constants'),
+    lastSeenItems: jest.fn().mockReturnValue([])
+
+}));
 
 describe('EntityTableToolbar', () => {
     let initialState;
@@ -92,6 +99,7 @@ describe('EntityTableToolbar', () => {
     });
 
     describe('DOM', () => {
+
         it('should render correctly - no data', () => {
             const store = mockStore({
                 entities: {
@@ -322,7 +330,7 @@ describe('EntityTableToolbar', () => {
                 </Provider>);
                 wrapper.find('.ins-c-chip-filters button.pf-m-link').last().simulate('click');
                 const actions = store.getActions();
-                expect(actions.length).toBe(3);
+                expect(actions.length).toBe(4);
                 expect(actions[actions.length - 2]).toMatchObject({ type: 'CLEAR_FILTERS' });
                 expect(onRefreshData).toHaveBeenCalledWith({ filters: [], page: 1 });
             });
@@ -347,7 +355,7 @@ describe('EntityTableToolbar', () => {
 
             const wrapper = mount(<Provider store={store}>
                 <EntityTableToolbar
-                    hideFilters={{ all: true, name: false }}
+                    hideFilters={{ all: true, name: false, group: true }}
                     page={1}
                     total={500}
                     perPage={50}
@@ -384,6 +392,17 @@ describe('EntityTableToolbar', () => {
             </Provider>);
             expect(toJson(wrapper.find('Reset Filter')));
             expect(toJson(wrapper.find('Test Reset Filter'))).toBeFalsy();
+        });
+    });
+
+    describe('System update method filter', () => {
+        it('Should hide the filter when flag is disabled', () => {
+            const store = mockStore(initialState);
+            const wrapper = mount(<Provider store={store}>
+                <EntityTableToolbar hasItems onRefreshData={onRefreshData} loaded
+                    activeFiltersConfig={{ deleteTitle: 'Test Reset Filters' }} />
+            </Provider>);
+            expect(toJson(wrapper.find('System Update Method'))).toBeFalsy();
         });
     });
 });