+++ src/components/InventoryTable/InventoryTable.test.js
@@ -12,6 +12,9 @@ import toJson from 'enzyme-to-json';
 import { ConditionalFilter } from '@redhat-cloud-services/frontend-components/ConditionalFilter';
 import * as loadSystems from '../../Utilities/sharedFunctions';
 import { mockSystemProfile } from '../../__mocks__/hostApi';
+import useFeatureFlag from '../../Utilities/useFeatureFlag';
+
+jest.mock('../../Utilities/useFeatureFlag');
 
 describe('InventoryTable', () => {
     let initialState;
@@ -208,6 +211,7 @@ describe('InventoryTable', () => {
         });
 
         it('should disable only one filter', () => {
+            useFeatureFlag.mockReturnValue(true);
             const store = mockStore(initialState);
             const wrapper = mount(<Provider store={ store }>
                 <Router>
@@ -216,7 +220,14 @@ describe('InventoryTable', () => {
             </Provider>);
 
             expect(wrapper.find(ConditionalFilter).props().items.map(({ label }) => label)).toEqual(
-                ['Status', 'Operating System', 'Data Collector', 'RHC status', 'System Update Method', 'Tags']
+                ['Status',
+                    'Operating System',
+                    'Data Collector',
+                    'RHC status',
+                    'System Update Method',
+                    'Last seen',
+                    'Group',
+                    'Tags']
             );
         });