+++ src/routes/InventoryTable.js
@@ -48,8 +48,13 @@ const filterMapper = {
         flatMap(tagFilters, mapTags)
     ),
     rhcdFilter: ({ rhcdFilter }, searchParams) => rhcdFilter?.forEach(item => searchParams.append(RHCD_FILTER_KEY, item)),
+    //TODO: Add a way to add to url in a way that allows shareable links
+    // lastSeenFilter: ({ lastSeenFilter }, searchParams) =>
+    //     Object.values(lastSeenFilter)?.forEach(item => searchParams.append('last_seen', item)),
     updateMethodFilter: ({ updateMethodFilter }, searchParams) =>
-        updateMethodFilter?.forEach(item => searchParams.append(UPDATE_METHOD_KEY, item))
+        updateMethodFilter?.forEach(item => searchParams.append(UPDATE_METHOD_KEY, item)),
+    groupHostFilter: ({ groupHostFilter }, searchParams) => groupHostFilter
+    ?.forEach(item => searchParams.append('host_group', item))
 };
 
 const calculateFilters = (searchParams, filters = []) => {
@@ -58,7 +63,6 @@ const calculateFilters = (searchParams, filters = []) => {
             filterMapper?.[key]?.(filter, searchParams);
         });
     });
-
     return searchParams;
 };
 
@@ -78,10 +82,12 @@ const Inventory = ({
     operatingSystem,
     rhcdFilter,
     updateMethodFilter,
+    lastSeenFilter,
     page,
     perPage,
     initialLoading,
-    hasAccess
+    hasAccess,
+    groupHostsFilter
 }) => {
     const history = useHistory();
     const chrome = useChrome();
@@ -89,7 +95,16 @@ const Inventory = ({
     const [isModalOpen, handleModalToggle] = useState(false);
     const [currentSytem, activateSystem] = useState({});
     const [filters, onSetfilters] = useState(
-        generateFilter(status, source, tagsFilter, filterbyName, operatingSystem, rhcdFilter, updateMethodFilter)
+        generateFilter(
+            status,
+            source,
+            tagsFilter,
+            filterbyName,
+            operatingSystem,
+            rhcdFilter,
+            updateMethodFilter,
+            groupHostsFilter,
+            lastSeenFilter)
     );
     const [ediOpen, onEditOpen] = useState(false);
     const [globalFilter, setGlobalFilter] = useState();
@@ -289,7 +304,9 @@ Inventory.propTypes = {
     initialLoading: PropTypes.bool,
     rhcdFilter: PropTypes.oneOfType([PropTypes.arrayOf(PropTypes.string), PropTypes.string]),
     updateMethodFilter: PropTypes.oneOfType([PropTypes.arrayOf(PropTypes.string), PropTypes.string]),
-    hasAccess: PropTypes.bool
+    hasAccess: PropTypes.bool,
+    groupHostsFilter: PropTypes.oneOfType([PropTypes.arrayOf(PropTypes.string), PropTypes.string]),
+    lastSeenFilter: PropTypes.oneOfType([PropTypes.arrayOf(PropTypes.string), PropTypes.string])
 };
 
 Inventory.defaultProps = {