+++ src/store/inventory-actions.js
@@ -3,7 +3,6 @@ import {
     SELECT_ENTITY,
     CHANGE_SORT,
     FILTER_ENTITIES,
-    APPLICATION_SELECTED,
     FILTER_SELECT,
     UPDATE_ENTITIES,
     ENTITIES_LOADING,
@@ -21,6 +20,7 @@ import {
     filtersReducer,
     getOperatingSystems
 } from '../api';
+import { getGroupDetail, getGroups } from '../components/InventoryGroups/utils/api';
 
 export const loadEntities = (items = [], { filters, ...config }, { showTags } = {}, getEntities = defaultGetEntities) => {
     const itemIds = items.reduce((acc, curr) => (
@@ -37,11 +37,14 @@ export const loadEntities = (items = [], { filters, ...config }, { showTags } = 
         ...filters.length === 0 && { registeredWithFilter: [] },
         ...(isFilterDisabled('stale') && { staleFilter: undefined }),
         ...(isFilterDisabled('registeredWith') && { registeredWithFilter: undefined }),
-        ...(isFilterDisabled('operating_system') && { osFilter: undefined })
+        ...(isFilterDisabled('operating_system') && { osFilter: undefined }),
+        ...(isFilterDisabled('host_group')) && { groupHostFilter: undefined }
     }) : {
         ...(isFilterDisabled('stale') && { staleFilter: undefined }),
+        ...(isFilterDisabled('last_seen') && { lastSeenFilter: undefined }),
         ...(isFilterDisabled('registeredWith') && { registeredWithFilter: undefined }),
-        ...(isFilterDisabled('operating_system') && { osFilter: undefined })
+        ...(isFilterDisabled('operating_system') && { osFilter: undefined }),
+        ...(isFilterDisabled('host_group')) && { groupHostFilter: undefined }
     };
 
     const orderBy = config.orderBy || 'updated';
@@ -111,11 +114,6 @@ export const filterEntities = (key, filterString) => ({
     payload: { key, filterString }
 });
 
-export const detailSelect = (appName) => ({
-    type: APPLICATION_SELECTED,
-    payload: { appName }
-});
-
 export const entitiesLoading = (isLoading = true) => ({
     type: ENTITIES_LOADING,
     payload: { isLoading }
@@ -176,6 +174,16 @@ export const fetchAllTags = (search, pagination, getTags = defaultGetAllTags) =>
     type: ACTION_TYPES.ALL_TAGS,
     payload: getTags(search, pagination),
     meta: { lastDateRequestTags: Date.now() }
+});
+
+export const fetchGroups = (search, pagination) => ({
+    type: ACTION_TYPES.GROUPS,
+    payload: getGroups(search, pagination)
+});
+
+export const fetchGroupDetail = (groupId) => ({
+    type: ACTION_TYPES.GROUP_DETAIL,
+    payload: getGroupDetail(groupId)
 });
 
 export const fetchOperatingSystems = (params = []) => ({