+++ src/store/reducers.js
@@ -11,6 +11,8 @@ import { mergeArraysByKey } from '@redhat-cloud-services/frontend-components-uti
 import { notificationsReducer } from '@redhat-cloud-services/frontend-components-notifications/redux';
 import entitiesReducer, { defaultState as entitiesDefault } from './entities';
 import entityDetailsReducer, { entityDefaultState as entityDefault, updateEntity } from './entityDetails';
+import groups from './groups';
+import groupDetail from './groupDetail';
 
 export { entitiesReducer, entityDetailsReducer };
 
@@ -112,7 +114,9 @@ function onSetPagination(state, { payload }) {
 
 let reducers = {
     notifications: notificationsReducer,
-    systemProfileStore
+    systemProfileStore,
+    groups,
+    groupDetail
 };
 
 export const tableReducer = applyReducerHash(